@@ -5,8 +5,7 @@
 #pragma warning(push)
 #pragma warning(disable : 6294) // Ill-defined for-loop
 
-namespace Ray {
-namespace NS {
+namespace Ray::NS {
 void ClearBorders(const rect_t &rect, const int w, const int h, const bool downscaled, const int out_channels,
                   float output[]) {
     if (!downscaled) {
@@ -586,7 +585,6 @@ void ConvolutionConcat3x3(const float *restrict data1, const float *restrict dat
 
 #undef index1
 }
-} // namespace NS
-} // namespace Ray
+} // namespace Ray::NS
 
 #pragma warning(pop)
