@@ -3,9 +3,7 @@
 #include "../RendererBase.h"
 #include "Dx/ContextDX.h"
 
-namespace Ray {
-namespace Dx {
+namespace Ray::Dx {
 RendererBase *CreateRenderer(const settings_t &s, ILog *log,
                              const std::function<void(int, int, ParallelForFunction &&)> &parallel_for);
-} // namespace Dx
-} // namespace Ray
+} // namespace Ray::Dx
