@@ -553,7 +553,7 @@ void Ray::Cpu::Scene::RemoveMesh_nolock(const MeshHandle i) {
 
     meshes_.Erase(i._block);
 
-    bool rebuild_required = false;
+    [[maybe_unused]] bool rebuild_required = false;
     for (auto it = mesh_instances_.begin(); it != mesh_instances_.end();) {
         mesh_instance_t &mi = *it;
         if (mi.mesh_index == i._index) {
@@ -563,7 +563,6 @@ void Ray::Cpu::Scene::RemoveMesh_nolock(const MeshHandle i) {
             ++it;
         }
     }
-    unused(rebuild_required);
 
     if (use_wide_bvh_) {
         mtris_.Erase(tris_block);
@@ -1193,7 +1192,7 @@ void Ray::Cpu::Scene::PrepareEnvMapQTree_nolock() {
     for (int i = 0; i < env_.qtree_levels; ++i) {
         env_.qtree_mips[i] = value_ptr(env_map_qtree_.mips[i][0]);
     }
-    for (int i = env_.qtree_levels; i < countof(env_.qtree_mips); ++i) {
+    for (int i = env_.qtree_levels; i < std::size(env_.qtree_mips); ++i) {
         env_.qtree_mips[i] = nullptr;
     }
 
@@ -1456,9 +1455,8 @@ void Ray::Cpu::Scene::RebuildLightTree_nolock() {
     }
 
     if (use_wide_bvh_ || true) {
-        const uint32_t root_node = FlattenLightBVH_r(light_nodes_, 0, light_cwnodes_);
+        [[maybe_unused]] const uint32_t root_node = FlattenLightBVH_r(light_nodes_, 0, light_cwnodes_);
         assert(root_node == 0);
-        unused(root_node);
         light_nodes_.clear();
 
         // Collapse leaf level (all leafs have only 1 light)
