@@ -4,8 +4,7 @@
 #include "TextureAtlasVK.h"
 #include "TextureVK.h"
 
-namespace Ray {
-namespace Vk {
+namespace Ray::Vk {
 const Bitmask<eStage> g_stage_bits_per_state[] = {
     {},                  // Undefined
     eStage::VertexInput, // VertexBuffer
@@ -90,7 +89,7 @@ const VkImageLayout g_image_layout_per_state_vk[] = {
     VK_IMAGE_LAYOUT_UNDEFINED,                        // BuildASWrite
     VK_IMAGE_LAYOUT_UNDEFINED                         // RayTracing
 };
-static_assert(COUNT_OF(g_image_layout_per_state_vk) == int(eResState::_Count), "!");
+static_assert(std::size(g_image_layout_per_state_vk) == int(eResState::_Count), "!");
 
 const VkAccessFlags g_access_flags_per_state_vk[] = {
     0,                                                                                          // Undefined
@@ -110,7 +109,7 @@ const VkAccessFlags g_access_flags_per_state_vk[] = {
     VK_ACCESS_ACCELERATION_STRUCTURE_READ_BIT_KHR | VK_ACCESS_ACCELERATION_STRUCTURE_WRITE_BIT_KHR, // BuildASWrite
     VK_ACCESS_SHADER_READ_BIT,                                                                      // RayTracing
 };
-static_assert(COUNT_OF(g_access_flags_per_state_vk) == int(eResState::_Count), "!");
+static_assert(std::size(g_access_flags_per_state_vk) == int(eResState::_Count), "!");
 
 const VkPipelineStageFlags g_pipeline_stages_per_state_vk[] = {
     {},                                 // Undefined
@@ -134,9 +133,8 @@ const VkPipelineStageFlags g_pipeline_stages_per_state_vk[] = {
     VK_PIPELINE_STAGE_ACCELERATION_STRUCTURE_BUILD_BIT_KHR, // BuildASWrite
     VK_PIPELINE_STAGE_RAY_TRACING_SHADER_BIT_KHR            // RayTracing
 };
-static_assert(COUNT_OF(g_pipeline_stages_per_state_vk) == int(eResState::_Count), "!");
-} // namespace Vk
-} // namespace Ray
+static_assert(std::size(g_pipeline_stages_per_state_vk) == int(eResState::_Count), "!");
+} // namespace Ray::Vk
 
 Ray::Bitmask<Ray::Vk::eStage> Ray::Vk::StagesForState(const eResState state) {
     return g_stage_bits_per_state[int(state)];
